// bookstore-webapp/src/generated/orders-service.ts
export interface components {
    schemas: {
        OrderItem: {
            bookId: string;
            quantity: number;
            unitPrice: number;
        };
        Order: {
            id: string;
            /** @enum {string} */
            status: "placed" | "confirmed" | "shipped" | "delivered" | "cancelled";
            total: number;
            addressId?: string;
            items: components["schemas"]["OrderItem"][];
            placedAt: string;
        };
        Error: {
            code: number;
            message: string;
            description?: string;
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}